Fix infinite loop while reporting an error (backport)
This is CVE-2017-9210.

The description string for an error message included unparsing an
object, which is too complex of a thing to try to do while throwing an
exception. There was only one example of this in the entire codebase,
so it is not a pervasive problem. Fixing this eliminated one class of
infinite loop errors.

(cherry picked from commit 603f222365252f1a1e20303b3dbe52466be3053b)