Perl wrongly writes individual UTF-8 bytes in test file names.
Those file names are needed to test Unicode compatibility for Windows and the problem can easily be fixed by creating the necessary file names manually and version them. That way they always have the correct encoding in the currently available file system, as GIT-clients should handle that properly. It's not possible to alter the name bytes in the Perl test using this commit, which is a strong indication that things are handled properly.