Exclude signature dictionary from compressible objects
It seems better not to compress signature dictionaries.
Various PDF digitally signing tools, including Adobe Acrobat Reader DC,
do not compress signature dictionaries.

Table 8.93 "Entries in a signature dictionary" in PDF 1.5 reference
describes that /ByteRange in the signature dictionary
shall be used to describe a digest
that does not include the signature value (/Contents) itself.

The byte ranges cannot be determined if the dictionary is compressed.