wip: avoid mutating stream dictionaries while writing
The code works for files in the test suite when not linearizing. There
could be bugs around pathological cases in which an object is
indirectly referenced from /Length, /Filter, or /DecodeParms in a
stream dictionary and also in some other place. Also, this may not
actually reduce the number of traversals since we call
fixDanglingReferences(true) at the beginning of this.