What is CI? ci.debian.net? Took me some time
I hope I am not the only one wondering what CI is. Finding that out was not a quick search. Do other people find it more quickly? And I am still not sure it is what this patch refers to. Do give a hint about that, or use the full term, or apply the patch, or something. Thank you.