Forward declare JSON_array for C++20 compatibility.
In C++20 the destructor for vector<T> is constexpr. This requires that
T not be an incomplete type where the destructor is called.

JSON_array's `vector<JSON> elements` field violates this because JSON
is still an incomplete type when JSON_array's destructor is defined.

To fix this forward declare JSON_array and only define it afer JSON's
definition ends (once it is a complete type).

No new tests since I'm not really sure how to test it and qpdf doesn't
officially support C++20 yet. If there's a regression I'll notice it on
my end, but I'll only check once per release until my company is a bit
farther along in our C++20 migration.