Refactor QPDF::Objects / JSONReactor interaction
Allow for the fact that when processing JSON input we cannot determine
whether references are dangling until the whole input has been processed.

We handle this by optimistically storing references in the object table.
When additional gens are encountered for the same object we store them
in the new map unconfirmed_objects. Once processing is complete we clean
up the object table and clear unconfirmed_objects.

New test cases are adapted from manual-qpdf-json.json etc.