Added prefix m to page ranges
The new prefix m can be used to specify a number as the maximum of
pages that should be used.
Before these changes specifying a range larger than the amount of
pages in the document would give an error.
When using qpdf to split large numbers of documents with varying page
numbers it was not possible to use the same parameters to always get
a set number of the first pages.

For example to get at most the first 5 pages one would use 1-m5.

This commit changes the regular expression to parse page ranges to
accept m or r as prefixes. In addition the CLI help as well as the
online documentation have been expanded. A very basic test was added aswell.