Refactor xref table reconstruction (Fixes #1362)
Split reconstruction into three passes - scanning of input for objects and
trailer, insertion of objects into the xref table, and loading the trailer.

This allows insertion to take place in the usual reverse order and removes
the need for a separate insertReconstructedXrefEntry method.

It also allows trailer to be tried from most recent to oldest.

Ignore any found trailers without /Root entry.